requirements.txt
Flask-SQLAlchemy
Flask
SQLAlchemy
python-dotenv